% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFW.R
\name{fw.check}
\alias{fw.check}
\title{Check the configurable object against the data frame input}
\usage{
fw.check(config, data)
}
\arguments{
\item{config}{Config list object}

\item{data}{Data Frame you are intending to write}
}
\description{
Check the configurable object against the data frame input
}
