% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw.make.R
\name{fw.make}
\alias{fw.make}
\title{Write to a file using a configurable fixed width format}
\usage{
fw.make(conf, data, file = FALSE)
}
\arguments{
\item{conf}{Config list object}

\item{data}{Data Frame you are intending to write}

\item{file}{Write to File? (FALSE or string to be the file name) (optional)}
}
\description{
You can learn more about package authoring with RStudio at:
}
\examples{
# Example Configuration List
conf = list(
  item1 = list(
    start = 1,
    end = 10,
    align = "l",
    padding = " ",
    required = TRUE
  ),
  item2 = list(
    start = 11,
    end = 20,
    align = "r",
    padding = 0,
    required = FALSE,
    default = " "
  )
)

# Example Data
data = data.frame(item1 = c("AAAA", "BBBB", "CCCC"), item2 = c(254, 2.25, 1.4))

fw.make(conf, data, file = FALSE)
}
\seealso{
\code{\link{fw.check}} for config validation function
}
