% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeFW.R
\name{write.fw}
\alias{write.fw}
\title{Write to a file using a configurable fixed width format}
\usage{
write.fw(conf, data, file = FALSE)
}
\arguments{
\item{conf}{Config list object}

\item{data}{Data Frame you are intending to write}

\item{file}{FALSE or character verctor of desired file name}
}
\description{
You can learn more about package authoring with RStudio at:
}
