% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw.check.R
\name{fw.check}
\alias{fw.check}
\title{Check the configurable object against the data frame input}
\usage{
fw.check(config, data)
}
\arguments{
\item{config}{Config list object}

\item{data}{Data Frame you are intending to write}
}
\description{
This functions call be called directly to test that the data.frame you want to convert matches the config
file you have set up. This function is called directly within fw.make.
}
